1.0.0 NEW STABLE VERSION
This is the first stable version of the code.

The code is divided in two parts: driver + SS solver
The *driver* is meant to be specific of the problem to solve.
The *SS solver* is generic.
In the driver the ss_init must be invoked before the call to ss_solve.
*ss_init* and *ss_solve* are the two main interface of the code.
All the rest is tuned in the default input file, read with a call
to ss_read_input.

A working example is in either ss_hm_bethe (multi-band Bethe lattice
model) and ss_hm_2d (2d-square lattice multi-band model).
Other drivers will be available as time goes.

The code features different methods to solve the self-consistency of
the SS problem.
The default one transforms the problem into a root finding for the
functional F(X),
where X = [lambda,Z,Ef] or [lambda,Z] if filling is set to 0, while
F = [n-Sz-1/2,Z-Z_old,n-filling].
This way of solving is interfaced to either hybrd (in MINPACK) thru
FSOLVE (hybrd interface in SciFortran) and broyden1 (interface to
broyden method approximating the gradient).
This method has been tested against old data and other method
so far with great success. It offer more stability and faster execution.

Old methods, dubbed GG and LF, are available:
gg_broyden
gg_fsolve
lf_solve

gg_xyz uses broyden/fsolve on lambda and iterate over Z for any fixed lambda
lf_solve iterates over lambda, solve fermion at fixed lambda + broyden/fsolve
for spins changing lambda, fix Z
These methods are included for backcompatibility but are working (or should be).

Output is written out to files, with extension .ss, at each iteration.
The used input, if any is named parameters.used
The output is save into parameters.restart which can be used to restart the calculation.
To repeat the calculation *.used --> *.restart

More informations are available inside the code as comments.