1.0.1 UPDATE
- Added a driver for the PCO using Wannier90 interface (needs W90 bands structure file).
- Fixed a problem with the evaluation of the Fermi level in lambda0. We now use fzero rather
than brent method, which although a bit slower ensures a better convergence.

> REMARK: it is important that the band structure passed as input to SS solver is centered
at the Fermi level. For it is useful to perform this action in the driver.
One way to automatically do that is to indicate the putative Fermi level (perhaps as xmu
in the input) and use it to re-center the band structure. We regularly subtract the local
part of the Hamiltonian but this is required to obtain the Fermi level in lambda0 as it
contains also the component other than proportional to sigma_0.

I will work on that. For the time being user is advised.