1.0.2 Update
Code updated.

Some improvements in the handling of the occupation and band structure.
Entire code is now using chemical potential (gran-canonical ensemble) as opposed
to Fermi level.
Introduced a number of Input/Output functions to retrieve internal quantities as
zeta, lambda, renormalized fermionic Hamiltonian and so on.
These can be used for post-processing reasons.

This code is compliant with the latest updates in DMFT_TOOLS which enable to push
zeta and ReSigma(0) to evaluation of the renormalized tight-binding Hamiltonian H(k).
This can be used for example to get renormalized band structure or fermionic Green's
functions (z-H*(k))^-1

Next: evaluate the SS Green's function G = G_f * Pi_s