1.1.1 BUG FIX
Testing interface to Fe-based compounds. Started with LaOFeAs.
Solved a couple of minor but severe bugs: real(8) instead of complex(8)
matrices used to store the results of the diagonalization eigh in
ss_get_lambda0. produced a wrong density, wrong lambda0, etc.

There is an issue with the I/O procedure: the order of the arrays/matrices
is not necessarily the same. Fix it as in ss_init.

For U=0 it takes quite some time to converge at the correct density.
1-loop (of Fsolve call) is about 0.800sec. Not bad...

For harder problem we can use parallel solution: .x./eigh --> .px./p_eigh
(need SCALAPACK).