Intermediate commit. Bug fix.
This is an intermediate commit while searching and fixing the bug which
prevent correct behavior for generic W90 Hamiltonian.

Meanwhile I spotted a bug (up to now irrelevant) related to the
re-ordering of the local part of the Hamiltonian Hloc. Hloc was actually
not re-ordered at all, leading to a possible wrong sum up of H(k) with its
local part in SOLVE_FERMION.
Unfortunately this was not *the* bug. so far if Nspin=1 the conventional
user order [Norb,Nspin,Nlat] coincides with SS order [Norb,Nlat,Nspin].