1.4.0 BUG FIX + UPDATE
The main bug of the code has been solved.
This was related to the wrong setup of the local part of the
renormalized hamiltonian.

Theory: The local *diagonal* part of the non-interacting Hamiltonian,
which couples to the density, is not involved in the pseudo-particle
transformation d-->O*f. Thus it does not enter in the expression for
the renormalized fermionic Hamiltonian O*.H.O.
To this end we extract a local part from the input Hamiltonian H(k).
H*(k) = H(k) - Hloc

BUG: we obtained Hloc as sum(H(k)/Nk, which corresponds to the
*diagonal* part only if no inter-orbital couplings are present.

FIX: obtain Hloc as sum(diagonal(H(k)))/Nk.

TODO:
-Testing from LaOFeAs are positive but still there is a
factor of about 1/3. This is a yet another possible bug.
-Further reduce computational by solving fermionic Hamiltonian
for one spin, i.e. 1:Niso. Solution will then be propagated to
the Ns=2*Niso parameters using spin_symmetry function.