BUG FIX
The bug preventing SS code to properly work has been fixed.
Shame on me: the bug was hidden in an incorrect way to evaluate
the one-body density matrix <f^+_a f_b>, where I exchanged the
indices of the eigenvector matrices!

<f^+_a f_b> = sum_d u*_ad u_bd n_Fermi(E_d)

b and a in the sum above were interchanged causing a bad evaluation
of the off-diagonal terms and ultimately of the Weiss field contributions
which indeed were correctly coded.

Tested against LdM code.
- lambda0 and 1st Weiss field equal to 6digits

Tested against SS results published in the PRL "la pietra" show
essentially identical results.

Note that:
> lambda0 needs abs(Weiss)
> lambda is initialized to lambda0

This guarantees that in the first loop the code converges very rapidly as
lambda cancels out with lambda0, as by definition.

Ready to import treatment of all terms of the Hamiltonian.