1.8.0 STABLE
This is a stable debugged version reproducing the published results
for the LaOFeAs.

The main driver is ss_DFT, which interfaces to Wannier 90 Hamiltonian.
Other problems can be derived from this in a straightforward way,
see for example Bethe or 2d square lattice cases.

The code supports MPI to tackle the k-dependence, so it is expected
to scale almost linearly. Although execution is already quite fast
for one processor and 5 orbitals.

This version feature a complete separation of the Hamiltonian H(k)
and deals separately with all the local and non-local terms.
The hybridization part (local, intra-cell, off-diagonal) should
anyway be tested.

Ready for production.