2.0.4 UPDATE
Previous version 2.0.3 relied on Fortran2008 intrinsic FindLoc, which is available starting from GNU gcc suite 9.0.0.
To avoid relying on such an advanced tool, we introduced a homemade version of findloc, named ss_findloc, placed
in SS_VARS_GLOBAL, for integers and character. Tested in the officina.

To check compatibility with older compilers.