2.1.0 UPDATED CODE
This version of the code has been extended to take care of the
possible presence of Spin-Orbit Coupling terms appearing in the
Hamiltonian.

More in general the Pseudo-Fermion solver renormalizes the spin-off-diagonal
elements in the Hamiltonian.
Indeed, the code required just few retouches.
These new terms renormalize similarly to the non-diagonal elements already
present in the code.

TEST TO BE DONE.
CODE TESTED FOR BETHE 2BANDS PROBLEM WITH N=1.84: PERFECT MATCH.