2.1.1 Update Code: self-consistent Lambda0 calculation
This update includes a (experimental) self-consistent evaluation
of the lambda0 parameter, as in PRB 86, 084104 (2012).
This is achieved using a new routine ss_update_lambda (SS_SOLVE_FERMIONS)
which is called in the ss_solve_xyz functions after fermions and spins
problems are solved to update lambda0 for the next iteration.

To control this feature we introduced a new input variable:

logical :: self_lam0

false by default.

NOT TESTED